Add tests for pyenv environments; refactor BaseVirtualEnvironment class for correctness
