+++ python_venv/const.py
@@ -14,12 +14,19 @@ DEV_SUFFIX = "-dev"
 
 MESSAGE_PREFIX = "==> "
 
+ENV_DIR_PLACEHOLDER = "<ENV_DIR>"
+
 ENV_TYPE_VENV = "venv"
 ENV_TYPE_PYENV = "pyenv"
 ENV_TYPE_CONDA = "conda"
 
 ENV_TYPES = [
     ENV_TYPE_VENV,
+    ENV_TYPE_PYENV,
+    ENV_TYPE_CONDA,
+]
+
+ENV_TYPES_NAMED = [
     ENV_TYPE_PYENV,
     ENV_TYPE_CONDA,
 ]
